requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
moviepy==1.0.3
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.0 